Move ShpTD sprite loading into Mods.Common.
