+++ OpenRA.Game/OpenRA.Game.csproj
@@ -279,7 +279,6 @@
     <Compile Include="FileFormats\PngLoader.cs" />
     <Compile Include="FileFormats\R8Reader.cs" />
     <Compile Include="FileFormats\ShpD2Reader.cs" />
-    <Compile Include="FileFormats\ShpReader.cs" />
     <Compile Include="FileFormats\ShpTSReader.cs" />
     <Compile Include="FileFormats\TmpRAReader.cs" />
     <Compile Include="FileFormats\TmpTDReader.cs" />
