+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -113,6 +113,7 @@
     <Compile Include="World\ResourceClaim.cs" />
     <Compile Include="World\ResourceClaimLayer.cs" />
     <Compile Include="World\SmudgeLayer.cs" />
+    <Compile Include="SpriteLoaders\ShpTDLoader.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
